backend/pyproject.toml
[tool.poetry]
name = "baqsy-backend"
version = "0.1.0"
description = "Baqsy System — Django backend (business audit SaaS)"
authors = ["Baqsy Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
django = "5.2"
django-fsm-2 = "4.2.4"
psycopg2-binary = "2.9.10"
django-storages = {version = "1.14.6", extras = ["s3"]}
boto3 = ">=1.35"
django-environ = "^0.11"
weasyprint = "68.1"
gunicorn = "23.0.0"
celery = "5.6.3"
django-celery-beat = "2.7.0"
redis = "5.3.0"
django-cors-headers = "4.6.0"
structlog = "25.5.0"
django-structlog = "10.0.0"
djangorestframework = "3.17.1"
djangorestframework-simplejwt = "5.5.1"
django-unfold = "0.90.0"
django-axes = "8.3.1"
django-admin-sortable2 = "2.2.8"
django-ckeditor-5 = "0.2.20"
openai = ">=1.60,<3"
Pillow = "^10.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-django = "^4.9"
pytest-cov = "^5.0"
pytest-mock = "^3.14"
factory-boy = "^3.3"
fakeredis = "^2.23"
ruff = "^0.6"
mypy = "^1.11"
django-debug-toolbar = "^4.4"

[tool.poetry.group.prod.dependencies]
gunicorn = "23.0.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "baqsy.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-ra --strict-markers"
markers = [
    "slow: marks tests as slow",
    "integration: integration tests requiring services",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "DJ"]
ignore = ["E501"]
